+++ R/heplot.R
@@ -26,6 +26,8 @@
 # -- replaced internal label.ellipse with separate function; added label.pos= argument
 #  22 Feb 2013
 # -- added ... to label.ellipse to be able to pass cex=
+#  02 Dec 2025
+# added `label.cex` arg to allow control of H/E label sizes (not vectorized)
 
 
 
@@ -335,46 +337,47 @@ heplot.mlm <-
 				mod,           # an mlm object
 				terms,         # vector of terms to plot H ellipses
 				hypotheses,    # list of linear hypotheses for which to plot H ellipses
-				term.labels=TRUE,  # TRUE, FALSE or a vector of term labels of length(terms)
-				hyp.labels=TRUE,   # as above for term.labels
-				err.label="Error",
-				label.pos=NULL,    # label positions: NULL or 0:4
-				variables=1:2,     # x,y variables for the plot [variable names or numbers]
-				error.ellipse=!add,
-				factor.means=!add,
-				grand.mean=!add,
-				remove.intercept=TRUE,
-				type=c("II", "III", "2", "3"),
-				idata=NULL,
-				idesign=NULL,
-				icontrasts=c("contr.sum", "contr.poly"),
-				imatrix=NULL,
-				iterm=NULL,
-				markH0=!is.null(iterm),
+				term.labels = TRUE,  # TRUE, FALSE or a vector of term labels of length(terms)
+				hyp.labels = TRUE,   # as above for term.labels
+				err.label = "Error",
+				label.pos = NULL,    # label positions: NULL or 0:4
+				label.cex = par("cex"),
+				variables = 1:2,     # x,y variables for the plot [variable names or numbers]
+				error.ellipse = !add,
+				factor.means = !add,
+				grand.mean = !add,
+				remove.intercept = TRUE,
+				type = c("II", "III", "2", "3"),
+				idata = NULL,
+				idesign = NULL,
+				icontrasts = c("contr.sum", "contr.poly"),
+				imatrix = NULL,
+				iterm = NULL,
+				markH0 = !is.null(iterm),
 				manova,        # an optional Anova.mlm object
-				size=c("evidence", "effect.size", "significance"),
-				level=0.68,
-				alpha=0.05,
-				segments=60,   # line segments in each ellipse
-				center.pch="+",   # doesn't have to be an argument
-				center.cex=2,
+				size = c("evidence", "effect.size", "significance"),
+				level = 0.68,
+				alpha = 0.05,
+				segments = 60,   # line segments in each ellipse
+				center.pch = "+",   # doesn't have to be an argument
+				center.cex = 2,
 				col=getOption("heplot.colors", c("red", "blue", "black", "darkgreen", "darkcyan","magenta", 
 				                                 "brown","darkgray")),
 				# colors for H matrices, E matrix
-				lty=2:1,
-				lwd=1:2,
-				fill=FALSE,         ## whether to draw filled ellipses (vectorized)
-				fill.alpha=0.3,     ## alpha transparency for filled ellipses
+				lty = 2:1,
+				lwd = 1:2,
+				fill = FALSE,         ## whether to draw filled ellipses (vectorized)
+				fill.alpha = 0.3,     ## alpha transparency for filled ellipses
 				xlab,
 				ylab,
-				main="",
+				main = "",
 				xlim,           # min/max for X (override internal min/max calc) 
 				ylim,
-				axes=TRUE,      # whether to draw the axes
+				axes = TRUE,      # whether to draw the axes
 				offset.axes,    # if specified, the proportion by which to expand the axes on each end (e.g., .05)
-				add=FALSE,      # add to existing plot?
-				verbose=FALSE,
-				warn.rank=FALSE,  
+				add = FALSE,      # add to existing plot?
+				verbose = FALSE,
+				warn.rank = FALSE,  
 				...) {
 	ell <- function(center, shape, radius) {
 		angles <- (0:segments)*2*pi/segments
@@ -387,26 +390,9 @@ heplot.mlm <-
 		order <- order(attr(Q, "pivot"))
 		t( c(center) + t( circle %*% Q[,order]))
 	}
-#	label.ellipse <- function(ellipse, label, col){
-#		if (cor(ellipse)[1,2] >= 0){
-#			index <- which.max(ellipse[,2])
-#			x <- ellipse[index, 1] + 0.5 * strwidth(label)  # was: "A"
-#			y <- ellipse[index, 2] + 0.5 *strheight("A")
-#			adj <- c(1, 0) 
-#		}
-#		else {
-#			index <- which.min(ellipse[,2])
-#			x <- ellipse[index, 1] - 0.5 * strwidth(label)  # was: "A"
-#			y <- ellipse[index, 2] - 0.5 * strheight("A")
-#			adj <- c(0, 1) 
-#		}
-#		text(x, y, label, adj=adj, xpd=TRUE, col=col, ...)
-#	}
+
 #	last <- function(x) {x[length(x)]}
 	
-	#if (!require(car)) stop("car package is required.")
-	# avoid deprecated warnings from car
-#	if (packageDescription("car")[["Version"]] >= 2) linear.hypothesis <- linearHypothesis
 	type <- match.arg(type)
 	size <- match.arg(size)
 	data <- model.frame(mod)
@@ -485,11 +471,13 @@ heplot.mlm <-
 	col <- he.rep(col, n.ell) 
 	lty <- he.rep(lty, n.ell)
 	lwd <- he.rep(lwd, n.ell)
+	
 	# handle filled ellipses
 	fill <- he.rep(fill, n.ell)
 	fill.alpha <- he.rep(fill.alpha, n.ell)
 	fill.col <- trans.colors(col, fill.alpha)
 	label.pos <- he.rep(label.pos, n.ell)
+	
 	# TODO:  take account of rank=1?
 	fill.col <- ifelse(fill, fill.col, NA)
 	E.col<- last(col)
@@ -515,7 +503,8 @@ heplot.mlm <-
 			}
 			H.ellipse[[term]] <- ell(gmean, H, radius)
 			H.rank[term] <- qr(H)$rank
-		}
+	}
+	
 	if (n.hyp > 0) for (hyp in 1:n.hyp){
 			lh <- car::linearHypothesis(mod, hypotheses[[hyp]])
 			H <- lh$SSPH[variables, variables]
@@ -527,12 +516,14 @@ heplot.mlm <-
 				print(lh)
 			}
 			H.ellipse[[n.terms + hyp]] <- ell(gmean, H, radius)
-		}
+	}
+	
 	E <- SSPE
 	E <- E[variables, variables]
 	E <- E * scale[1]
 	E.ellipse <- ell(gmean, E, radius)
-	H.ellipse$E <- E.ellipse     
+	H.ellipse$E <- E.ellipse    
+	
 	if (!add){
 		max <- apply(sapply(H.ellipse, function(X) apply(X, 2, max)), 1, max)
 		min <- apply(sapply(H.ellipse, function(X) apply(X, 2, min)), 1, min)
@@ -565,31 +556,33 @@ heplot.mlm <-
 	}
 	# no longer need H.ellipse$E, since we return it separately
 	H.ellipse$E <- NULL
+	
 	if (grand.mean) 
 		points(gmean[1], gmean[2], pch=center.pch, cex=center.cex, col="black", xpd=TRUE)
+
 	if (error.ellipse){
-#		lines(E.ellipse, col=E.col, lty=lty[length(lty)], lwd=lwd[length(lwd)])
 		polygon(E.ellipse, col=last(fill.col), border=last(col), lty=last(lty), lwd=last(lwd))
-		label.ellipse(E.ellipse, err.label, col=last(col), label.pos=last(label.pos), ...)
+		label.ellipse(E.ellipse, err.label, col=last(col), label.pos=last(label.pos), cex=label.cex, ...)
 	}
 	term.labels <- if (n.terms == 0) NULL
 			else if (!is.logical(term.labels)) term.labels
 			else if (term.labels) terms else rep("", n.terms)  
 	if (n.terms > 0) for (term in 1:n.terms){
-#			lines(H.ellipse[[term]], col=col[term], lty=lty[term], lwd=lwd[term])
 			# TODO: avoid polygon if rank=1 ???
 			polygon(H.ellipse[[term]], col=fill.col[term], border=col[term],  lty=lty[term], lwd=lwd[term])
-			label.ellipse(H.ellipse[[term]], term.labels[term], col=col[term], label.pos=label.pos[term], ...) 
+			label.ellipse(H.ellipse[[term]], term.labels[term], col=col[term], label.pos=label.pos[term], cex=label.cex, ...) 
 		}   
+
+	# draw and label the H ellipses for all terms
 	hyp.labels <- if (n.hyp == 0) NULL
 			else if (!is.logical(hyp.labels)) hyp.labels
 			else if (hyp.labels) names(hypotheses) else rep("", n.hyp)  
 	if (n.hyp > 0) for (hyp in 1:n.hyp){
 			ell <- n.terms + hyp
-#			lines(H.ellipse[[ell]], col=col[ell], lty=lty[ell], lwd=lwd[ell])
 			polygon(H.ellipse[[ell]], col=fill.col[ell], border=col[ell],  lty=lty[ell], lwd=lwd[ell])
-			label.ellipse(H.ellipse[[ell]], hyp.labels[hyp], col=col[ell], label.pos=label.pos[hyp], ...)
+			label.ellipse(H.ellipse[[ell]], hyp.labels[hyp], col=col[ell], label.pos=label.pos[hyp], cex=label.cex, ...)
 		}
+
 	if (!add && (!is.logical(factor.means) || factor.means)){
 		for (fac in factors){
 			means <- aggregate(Y, list(fac), mean)
@@ -598,12 +591,22 @@ heplot.mlm <-
 		}
 	}
 	
+	# mark the null hypothesis, where relevant
 	if(is.logical(markH0) && markH0) mark.H0()
 	else if (is.list(markH0)) do.call(mark.H0, markH0)
 	
+	# prepare returned `"heplot"` result
 	names(H.ellipse) <- c(if (n.terms > 0) term.labels, if (n.hyp > 0) hyp.labels)
-	result <- if (!add) list(H=H.ellipse, E=E.ellipse, center=gmean, xlim=xlim, ylim=ylim, radius=radius)
-			else list(H=H.ellipse, E=E.ellipse, center=gmean, radius=radius)
+	result <- if (!add) list(H=H.ellipse, 
+	                         E=E.ellipse, 
+	                         center=gmean, 
+	                         xlim=xlim, 
+	                         ylim=ylim, 
+	                         radius=radius)
+			else list(H=H.ellipse, 
+			          E=E.ellipse, 
+			          center=gmean, 
+			          radius=radius)
 	class(result) <- "heplot"
 	invisible(result)
 }